update shared code

+++ smartmontools-arm-musl.sh
@@ -918,17 +918,21 @@ finalize_build() {
 
 # temporarily hide shared libraries (.so) to force cmake to use the static ones (.a)
 hide_shared_libraries() {
-    mv -f "${PREFIX}/lib_hidden/"* "${PREFIX}/lib/" || true
+    if [ -d "${PREFIX}/lib_hidden" ]; then
+        mv -f "${PREFIX}/lib_hidden/"* "${PREFIX}/lib/" || true
+        rmdir "${PREFIX}/lib_hidden" || true
+    fi
     mkdir -p "${PREFIX}/lib_hidden" || true
     mv -f "${PREFIX}/lib/"*".so"* "${PREFIX}/lib_hidden/" || true
-    mv -f "${PREFIX}/lib_hidden/libcc1."* "${PREFIX}/lib/" || true
     return 0
 }
 
 # restore the hidden shared libraries
 restore_shared_libraries() {
-    mv -f "${PREFIX}/lib_hidden/"* "${PREFIX}/lib/" || true
-    rmdir "${PREFIX}/lib_hidden" || true
+    if [ -d "${PREFIX}/lib_hidden" ]; then
+        mv -f "${PREFIX}/lib_hidden/"* "${PREFIX}/lib/" || true
+        rmdir "${PREFIX}/lib_hidden" || true
+    fi
     return 0
 }
 
@@ -1093,11 +1097,14 @@ if [ ! -f "$PKG_SOURCE_SUBDIR/__package_installed" ]; then
     unpack_archive "$PKG_SOURCE" "$PKG_SOURCE_SUBDIR"
     cd "$PKG_SOURCE_SUBDIR"
 
+    export LDFLAGS="-static ${LDFLAGS}" # use static linking for tests run by configure
+
     ./configure \
-         LDFLAGS="-static" \
          --prefix="${PREFIX}" \
          --host="${HOST}" \
     || handle_configure_error $?
+
+    export LDFLAGS="-all-static ${LDFLAGS}" # make static executable
 
     $MAKE
     make install
